Give CarPlay the quality the user actually chose
The streaming-quality rule — WiFi setting on WiFi, cellular setting on
cellular, 'high' when the connection is unknown — was derived inside
PlayingContext and nowhere else. CarPlay is the only other place that builds
a stream URL, and it asked for 'high' unconditionally: choose Original on
WiFi and the phone streamed lossless while the car got 320kbps mp3, with
nothing in the UI to explain the difference. The Opus setting was ignored
there too.

The rule moves to useStreamQuality so there is one answer to "what should we
ask for right now" and both callers read it.
